+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 1.1.0
 ### Added
 - Added fix for Flux Rifts not collapsing when fully consumed by Void Siphons
+- Added tweak increasing vertical search range when the research table scans for nearby research aids
 
 ### Changed
 ### Fixed